Split tz stuff in own module (issue 2)
