+++ app.py
@@ -1,18 +1,12 @@
 #!python3
-#This is the application script for launching a Python Timezone List Flask Web App.
+# This is the application script for launching
+# a Python Timezone List Flask Web App.
 
 from flask import Flask, render_template, request
-import pendulum
-import pytz
+
+from timezones import create_list, get_tz_time
 
 app = Flask(__name__)
-
-#Create the list to populate the pull-down menu
-def create_list():
-    tz_list = []
-    for tz in pytz.all_timezones:
-        tz_list.append(tz)
-    return tz_list
 
 
 @app.route('/', methods=['GET', 'POST'])
@@ -22,13 +16,12 @@ def index():
     tz_time = ''
     if request.method == 'POST' and 'tz_menu' in request.form:
         choice = request.form.get('tz_menu')
-        tz_time = pendulum.now(choice).to_datetime_string()
+        tz_time = get_tz_time(choice)
     return render_template('index.html',
-                            tz_list=tz_list,
-                            choice=choice,
-                            tz_time=tz_time)
+                           tz_list=tz_list,
+                           choice=choice,
+                           tz_time=tz_time)
 
-    
+
 if __name__ == "__main__":
     app.run()
-
